# openapi.yaml
openapi: 3.0.0
info:
  title: Java Faktury Project API
  description: "API for project about generating invoices using Java, Spring Boot,\
    \ MongoDB and React"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://faktury-backend.herokuapp.com/api
  description: API Server
security:
- bearerAuth: []
...
paths:
...
  /product:
    get:
      tags:
      - product
      summary: |
        search for existing products - retrieves all if searchstr blank (length = 0 or just a space)
      operationId: getProducts
      parameters:
      - name: searchstr
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: straws
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "500":
          description: internal server error
    post:
      tags:
      - product
      summary: add new product to database
      operationId: newProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "201":
          description: product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_201'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "409":
          description: product already exists
        "500":
          description: internal server error
  /product/{id}:
    get:
      tags:
      - product
      summary: get product data with ID
      operationId: getProductWithId
      parameters:
      - name: id
        in: path
        description: ID of the product to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: product not found
        "500":
          description: internal server error
    put:
      tags:
      - product
      summary: update product data
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        description: ID of the product to update
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such product
        "500":
          description: internal server error
    delete:
      tags:
      - product
      summary: delete specified product
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        description: ID of the product to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such product
        "500":
          description: internal server error
components:
  schemas:
...
    Product:
      required:
      - currency
      - name
      - price
      - tax
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: straws
        price:
          type: number
          example: 35.9
        tax:
          type: integer
          example: 23
        currency:
          type: string
          example: PLN
        unit:
          type: string
          example: szt.
...
    inline_response_201:
      type: object
      properties:
        id:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT